Funnel data analysis
A funnel is a marketing model which illustrates the theoretical customer journey towards the purchase of a product or service. Oftentimes, we want to track how many users complete a series of steps and know which steps have the most number of users giving up.

Some examples include:

Answering each part of a 5 question survey on customer satisfaction
Clicking “Continue” on each step of a set of 5 onboarding modals
Browsing a selection of products → Viewing a shopping cart → Making a purchase
Generally, we want to know the total number of users in each step of the funnel, as well as the percent of users who complete each step.