Calculating Churn Review
The churn rate is a percent of subscribers at the beginning of a period that cancel within that period. “Monthly churn” is a typical metric and what was used in the examples.
How to calculate this metric using SQL for a single month. This used COUNT() and conditions to determine the number of subscribers that were active and how many canceled.
A more complex method to track the subscriber churn rate over many months.